requirements.txt
pyjwt==2.4.0
cryptography>=3.3.1
jsonschema==4.15.0
boto3==1.24.66
Pillow==9.2.0
requests==2.28.1
shortuuid==1.0.9
pyIIIFpres==0.1.1
python-magic==0.4.27
Deprecated==1.2.13
backports.zoneinfo>=0.2.1;python_version<"3.9"